2.3-rc.3

+++ src/main/java/me/refracdevelopment/simplegems/commands/command/PayCommand.java
@@ -20,7 +20,7 @@ public class PayCommand extends RoseCommand {
     }
 
     @RoseExecutable
-    public void execute(CommandContext context, OfflinePlayer target, long amount, @Optional String silent) {
+    public void execute(CommandContext context, OfflinePlayer target, double amount, @Optional String silent) {
         final LocaleManager locale = this.rosePlugin.getManager(LocaleManager.class);
 
         if (context.getArgs()[1].contains("-")) return;
