+++ src/main/java/me/refracdevelopment/simplegems/data/Stat.java
@@ -2,25 +2,25 @@ package me.refracdevelopment.simplegems.data;
 
 public class Stat {
 
-    private long amount;
+    private double amount;
 
-    public long getAmount() {
+    public double getAmount() {
         return this.amount;
     }
 
-    public void setAmount(long amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 
-    public void incrementAmount(long amount) {
+    public void incrementAmount(double amount) {
         this.amount += amount;
     }
 
-    public void decrementAmount(long amount) {
+    public void decrementAmount(double amount) {
         this.amount -= amount;
     }
 
-    public boolean hasAmount(long amount) {
+    public boolean hasAmount(double amount) {
         return this.amount >= amount;
     }
 }
